Massive change in sight.
This commit is changing how the backend of the app is meant to work. Basically,
instead of relying on session.py to do all the grunt work, we're now calling on
connection.py to talk the Rest API ourselves. Most of the logic for each
command has been spread out in each classes instead of being centralised in a
single file. To add some flexibility to the mix, a new file has been
introduced which is bug.py, containing the Bug class. A Bug is the object
representation of a ... bug report but wrapped in a class. The logic in charge
of managing bug reports prior to this commit is very dirty and this overhaul
makes much more sense now.