reduce bug header CANDIDATES down to one
As far as Gentoo is concerned, there's only one candidate picked up by
libzilla now and that is the "Bug:" header. "Closes:" will be
automatically picked up by the Gentoo Bugzilla backend which, as its
name suggests, will close the bug in question. It might be interesting
to pick up the "Bug:" in case we want to further comment a bug from the
CLI. See GLEP 0066 for more information about all this.